# api/swagger/swagger.yaml
swagger: "2.0"
info:
  version: "0.0.1"
  title: Multiplayer Time API
# during dev, should point to your local machine
host: localhost:2000
...
basePath: /api
...
schemes:
  # tip: remove http to make production-grade
  - http
  - https
# format of bodies a client can send (Content-Type)
consumes:
  - application/json
# format of the responses to the client (Accepts)
produces:
  - application/json
...
securityDefinitions:
  jsonWebToken:
    type: apiKey
    in: header
    name: Authorization
  jsonAdminWebToken:
    type: apiKey
    in: header
    name: Authorization
paths:
...
  /checklogin:
    x-swagger-router-controller: authentication_controller
    post:
      operationId: getLogin
      tags: 
        - authentication
      description: Log in user in api
      parameters:
        - name: title
          description: User properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Token"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    get:
      operationId: checklogin
      tags: 
        - authentication
      security:
        - jsonWebToken: []
      description: Log in user in api
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/LoginResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  ################### Login #######################
  ################### Friends #####################
  
  ##TODODO
...
  /category:
    x-swagger-router-controller: category_controller
    get:
      operationId: getCategories
      tags:
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: get all categories
      responses:
        "200":
          description: Success
          schema:
              type: array
              items: 
                type: object
                allOf:
                  - $ref: "#/definitions/Category"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    post:
      operationId: addCategory
      tags: 
        - game
      security:
        - jsonAdminWebToken: []
      description: add a category
      parameters:
        - name: title
          description: Catergory properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteManyCategories
      tags:
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: remove many categories
      parameters:
        - in: body
          description: categories id's
          required: true
          name: categories
          schema:
            $ref: "#/definitions/ArrayId"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
  /category/{id}:
    x-swagger-router-controller: category_controller
    get:
      operationId: getCategory
      tags:
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: Get category by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/Category"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"  
    put:
      operationId: updateCategory
      tags: 
        - game
      security:
        - jsonAdminWebToken: []
      description: add a category
      parameters:
        - name: id
          type: string
          in: path
          required: true
        - name: category
          description: Catergory properties
          in: body
          required: true
          schema:
            $ref: "#/definitions/Category"
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse"
    delete:
      operationId: deleteCategory
      tags: 
        - game
      #security:
      #  - jsonAdminWebToken: []
      description: delete category by id
      parameters:
        - name: id
          type: string
          in: path
          required: true
      responses:
        "200":
          description: Success
          schema:
            $ref: "#/definitions/ApiResponse"
        default:
          description: Error
          schema:
            $ref: "#/definitions/ErrorResponse" 
            
  ################### CATEGORY ####################
  
  ################### GAMES #######################
...
definitions:
...
  ErrorResponse:
    required:
      - message
    properties:
      message:
        type: string
  ApiResponse:
    type: object
    properties:
      success:
        type: boolean
        description: returns true if successful
      description:
        type: string
        description: a short comment 
    required:
      - success
      - description
...
  LoginResponse:
    type: object
    properties:
      success:
        type: boolean
      messages:
        type: array
        items:
          type: object
          properties:
            id:
              type: string
            message:
              type: string
      token: 
        type: string
      username:
        type: string
      avatar:
        type: string
      country:
        type: string
      id:
        type: string
  Category:
    type: object
    properties:
      id:
        type: string
      name: 
        type: string 
      description: 
        type: string 
      color:
        type: string
...
  ArrayId:
    type: object
    properties:
      id:
        type: array
        items:
            type: string
...
  Token:
    type: object
    properties:
      token:
        type: string